requirements.txt
langgraph==0.2.59
langchain-openai==0.2.12
langchain-community==0.3.11
langchain_openai==0.2.12

azure-functions==1.18.0
python-dotenv==1.0.1
azure-storage-blob==12.19.1
azure-ai-formrecognizer==3.3.2
azure-ai-documentintelligence==1.0.0b2
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
azure-search-documents==11.5.2
azure-keyvault-secrets==4.8.0

tavily-python==0.5.0